cleanup and re-org
