+++ apps/browser/src/autofill/overlay/pages/button/autofill-overlay-button.ts
@@ -1,8 +1,8 @@
 import "@webcomponents/custom-elements";
 import "lit/polyfill-support.js";
 import { AuthenticationStatus } from "@bitwarden/common/auth/enums/authentication-status";
+import { EVENTS } from "@bitwarden/common/autofill/constants";
 
-import { EVENTS } from "../../../constants";
 import { buildSvgDomElement } from "../../../utils";
 import { logoIcon, logoLockedIcon } from "../../../utils/svg-icons";
 import {
