+++ apps/browser/src/autofill/overlay/pages/list/autofill-overlay-list.ts
@@ -1,9 +1,9 @@
 import "@webcomponents/custom-elements";
 import "lit/polyfill-support.js";
 import { AuthenticationStatus } from "@bitwarden/common/auth/enums/authentication-status";
+import { EVENTS } from "@bitwarden/common/autofill/constants";
 
 import { OverlayCipherData } from "../../../background/abstractions/overlay.background";
-import { EVENTS } from "../../../constants";
 import { buildSvgDomElement } from "../../../utils";
 import { globeIcon, lockIcon, plusIcon, viewCipherIcon } from "../../../utils/svg-icons";
 import {
