+++ apps/browser/src/autofill/overlay/pages/shared/autofill-overlay-page-element.ts
@@ -1,4 +1,5 @@
-import { EVENTS } from "../../../constants";
+import { EVENTS } from "@bitwarden/common/autofill/constants";
+
 import { RedirectFocusDirection } from "../../../utils/autofill-overlay.enum";
 import {
   AutofillOverlayPageElementWindowMessage,
