+++ apps/browser/tsconfig.json
@@ -31,5 +31,9 @@
     "strictTemplates": true,
     "preserveWhitespaces": true
   },
-  "include": ["src", "../../libs/common/src/platform/services/**/*.worker.ts"]
+  "include": [
+    "src",
+    "../../libs/common/src/platform/services/**/*.worker.ts",
+    "../../libs/common/src/autofill/constants"
+  ]
 }
