+++ libs/common/src/autofill/types/index.ts
@@ -1,0 +1,7 @@
+import { ClearClipboardDelay, AutofillOverlayVisibility } from "../constants";
+
+export type ClearClipboardDelaySetting =
+  (typeof ClearClipboardDelay)[keyof typeof ClearClipboardDelay];
+
+export type InlineMenuVisibilitySetting =
+  (typeof AutofillOverlayVisibility)[keyof typeof AutofillOverlayVisibility];
